--- Flask Test/templates/FunnyRandomText.html
@@ -1,7 +1,0 @@
-<link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='site.css')}}" />
-<html>
-	<body class="background">
-		<a href={{ url_for("home")}} class="backButton">Back</a>
-		<h1 class="headerOne">{{Country}}</h1>	
-	</body>
-</html>
