--- Flask Test/templates/LinkPage.html
@@ -1,10 +1,0 @@
-<link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='site.css')}}" />
-<script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.3/jquery.min.js"></script>
-<body class="background">
-    {{Buttons | safe }}
-    <br>
-    <a href={{url_for("Country")}}> Go To  Selected Data</a>
-    <div class="infoButton">
-        <a href={{ url_for("info")}}>Info</a>
-    </div>
-</body>
