+++ templates/Example Data
@@ -1,0 +1,6 @@
+0 1
+1 2
+2 6
+3 5
+4 3
+5 8
