+++ templates/info.html
@@ -1,0 +1,29 @@
+<link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='site.css')}}" />
+<body class="background">
+    <a href={{ url_for("home")}} class="backButton">Back</a>
+    <!--Devolper List-->
+    <h1 class="headerOne">Devolpers</h1>
+    <div class="sourceList">
+        <p style="color:#6dcefc">Brooke Malecki - Devolper</p>
+        <p style="color:#efa8b7">Vivian Brockman -Scrum Master</p>
+        <p style="color:white">Blake Darby -Product Owner</p>
+        <p style="color:#efa8b7">Reilly Robinson -Devolper</p>
+        <p style="color:#6dcefc">Sujay Baskiyar -Devolper</p>
+    </div>
+    <!--Copyright Information for the Data with hyperlinks to the source of the data-->
+    <h1 class="headerOne" style="border:6px solid #bc3b14;">Data Sources</h1>
+    <div class="sourceList">
+        <p>Climate Change: Earth Surface Temperature Data - Berkeley Earth 
+            <br>            <a href={{ url_for('DataSource1')}}>Link to the kaggle site where we got the data</a></p>
+    </div>
+    <!--Copyright Information for the images with hyperlinks to the sources of the images-->
+    <h1 class="headerOne" style="border:6px solid #797e1e;">Image Credits</h1>
+    <div class="sourceList">
+        <p>Title of the Image and where it is found- <a href={{ url_for('ImageS1')}}>Title of the Image</a></p>
+        <p>            (If Needed) Copyright Information</p>
+        <p>Title of the Image and where it is found- <a href={{ url_for('ImageS2')}}>Title of the Image</a></p>
+    </div>
+    <div class="infoButton">
+        <a href={{ url_for("home")}}>Info</a>   
+    </div>
+</body>
